+++ app/src/androidTest/java/com/learntodroid/booksearchr/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.learntodroid.mvvmrestapi;
+package com.learntodroid.booksearchr;
 
 import android.content.Context;
